add explicit cast for duration to floating point

+++ src/cw_tracer/cw_tracer.cpp
@@ -52,7 +52,7 @@ void cw_trace_span::resolve(size_t expected_id, ordered_json&& r) {
 double cw_trace_span::duration_us() const {
     cw_assert(resolved());
     auto elapsed_ns = std::chrono::duration_cast<std::chrono::nanoseconds>(end.value() - start).count();
-    return elapsed_ns / 1000.0;
+    return static_cast<double>(elapsed_ns) / 1000.0;
 }
 
 // ############### cw_tracer ###############
